Re-impl the `impl.client.rate-limiter` ns logic
After contacting the Telegram Bot Support, it's confirmed, in particular, that:
a. Each Telegram Bot API method has its own limits;
b. Limits for different API methods can vary significantly;
c. The Bot FAQ page only lists limits for the "sendMessage" method;
d. As an exception, the "editMessageText" shares limits with the "sendMessage"
   (and affects them too);
e. If a limit for an API method was reached, we need to retry the request later
   and also delay all the requests of the same type.

These details on the rate limiting strategy require a bit of a breaking change.